+++ inttest/codegeneration-defaults/model.cpp
@@ -7,7 +7,7 @@ namespace map_extras {
             try {
                 return *map_in.at(key_in);
             } catch (const std::out_of_range&) {
-                warning("No value found in param %s, ifmissing not specified", err.c_str());
+                warning("No value found in g3_param_table %s, ifmissing not specified", err.c_str());
                 return NAN;
             }
     }
@@ -56,7 +56,8 @@ Type objective_function<Type>::operator() () {
     PARAMETER(recage);
     PARAMETER(fish__Linf);
     PARAMETER(fish__K);
-    PARAMETER(fish__recl);
+    PARAMETER(fish__t0);
+    PARAMETER(fish__deltat);
     PARAMETER(fish__init__sd);
     PARAMETER(fish__walpha);
     PARAMETER(fish__wbeta);
@@ -363,7 +364,7 @@ Type objective_function<Type>::operator() () {
 
                     auto factor = (fish__init__scalar*map_extras::at_throw(fish__init, std::make_tuple(age), "fish.init")*exp(-(double)(1)*(fish__M + init__F)*(age - recage)));
 
-                    auto dnorm = ((fish__midlen - (fish__Linf*((double)(1) - exp(-(double)(1)*fish__K*(age - (recage + log((double)(1) - fish__recl / fish__Linf) / fish__K)))))) / fish__init__sd);
+                    auto dnorm = ((fish__midlen - (fish__Linf*((double)(1) - exp(-(double)(1)*fish__K*(age - (fish__t0 + fish__deltat)))))) / fish__init__sd);
 
                     {
                         fish__num.col(fish__age_idx).col(fish__area_idx) = normalize_vec(exp(-(pow(dnorm, (Type)(double)(2)))*(double)(0.5)))*(double)(10000)*factor;
@@ -670,7 +671,7 @@ Type objective_function<Type>::operator() () {
 
                     auto fish__area_idx = 0;
 
-                    auto dnorm = ((fish__midlen - (fish__Linf*((double)(1) - exp(-(double)(1)*fish__K*(age - (recage + log((double)(1) - fish__recl / fish__Linf) / fish__K)))))) / fish__rec__sd);
+                    auto dnorm = ((fish__midlen - (fish__Linf*((double)(1) - exp(-(double)(1)*fish__K*(age - (fish__t0 + fish__deltat)))))) / fish__rec__sd);
 
                     {
                         fish__renewalnum.col(fish__age_idx).col(fish__area_idx) = normalize_vec(exp(-(pow(dnorm, (Type)(double)(2)))*(double)(0.5)))*(double)(10000)*factor;